refactor: 예약 검증 로직 수정
